[MIG] mrp_bom_structure_report_level_1: Migration to 10.0
