+++ mrp_bom_structure_report_level_1/__manifest__.py
@@ -5,7 +5,7 @@
 
 {
     "name": "MRP BOM Structure Report Level 1",
-    "version": "9.0.1.0.0",
+    "version": "10.0.1.0.0",
     "author": "Eficent, Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/manufacture-reporting",
     "category": "Manufacture",
