+++ mrp_bom_structure_report_level_1/report/__init__.py
@@ -3,4 +3,4 @@
 #        (http://www.eficent.com)
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
-from . import bom_structure
+from . import mrp_bom_structure_report
