+++ mrp_bom_structure_report_level_1/static/description/index.html
@@ -67,7 +67,7 @@
         <div class="oe_span12">
             <h3>Contributors</h3>
             <ul>
-                <li>Miquel Raich &lt;<a
+                <li>Miquel Ra&iuml;ch &lt;<a
                         href="mailto:miquel.raich@eficent.com">miquel.raich@eficent.com</a>&gt;</li>
             </ul>
         </div>
