+++ mrp_bom_structure_report_level_1/views/report_mrpbomstructure.xml
@@ -1,6 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-<data>
+<odoo>
 <template id="report_mrp_bs_l1">
     <t t-call="report.html_container">
         <t t-call="report.internal_layout">
@@ -49,5 +48,4 @@
         </t>
     </t>
 </template>
-</data>
-</openerp>
+</odoo>
